Move MultiLineGridlikeTextTests into navigation package
